backend/requirements.txt
flask
pypdf2
flask-cors
google-generativeai
dotenv
flask-mail
cryptography
PyJWT